refactor: /application/apply 페이지 전환
